Only add bundles that aren't already present in the index image
Attempting to add bundles which are already present in the index
image causes an error. This patch seeks to prevent the task failure
it by filtering out the bundles which are already in the index
image. This solution had to be chosen because the API hasn't been
implemented and the error message offers no specific information.
This solution might significantly increase the IIB task runtime.